Add user account deletion functionality and related email templates

+++ backend/app/users/services.py
@@ -2,9 +2,21 @@ from typing import List
 
 from app.auth.repository import UserRepository
 from app.auth.schemas import UserResponse, UserRole
-from app.auth.security import get_password_hash, verify_password
+from app.auth.security import (
+    get_password_hash,
+    verify_password,
+    create_verification_token,
+    TokenType,
+    decode_verification_token,
+)
+from app.celery.tasks.email_tasks.tasks import (
+    user_deletion_confirmation,
+    user_deleted_notification,
+)
 from app.core.exceptions import AuthenticationError, NotFoundError
 from app.users.schemas import StudentData, UserUpdate
+from app.core.utils import make_username
+from app.settings import settings
 
 
 class UserService:
@@ -42,10 +54,47 @@ class UserService:
         updated_user = await self.user_repository.update(user_id, update_data)
         return UserResponse.model_validate(updated_user)
 
-    async def delete_user_info(self, user_id: str) -> None:
-        success = await self.user_repository.delete(user_id)
-        if not success:
+    async def request_delete_user_info(self, user_id: str) -> None:
+        user = await self.user_repository.get_by_id(user_id)
+        if not user:
             raise NotFoundError("User not found")
+
+        user_deletion_token = create_verification_token(
+            user_id=user_id, token_type=TokenType.USER_DELETION
+        )
+
+        user_deletion_link = settings.DELETE_ACCOUNT_URL.format(
+            token=user_deletion_token
+        )
+
+        user_deletion_confirmation.delay(
+            recipient=user.email,
+            link=user_deletion_link,
+            date_registered=user.created_at,
+            username=make_username(user),
+        )
+
+    async def delete_user_info(self, user_id: str, token: str) -> None:
+        data = decode_verification_token(token)
+        if not data or not data["type"]:
+            raise AuthenticationError(f"Invalid or expired token {data}")
+
+        if data["type"] != str(TokenType.USER_DELETION):
+            raise AuthenticationError("Invalid token type")
+
+        if data["user_id"] != user_id:
+            raise AuthenticationError("Invalid token user ID")
+
+        user = await self.user_repository.get_by_id(user_id)
+        if not user:
+            raise NotFoundError("User not found")
+
+        await self.user_repository.delete(user_id)
+        user_deleted_notification.delay(
+            recipient=user.email,
+            date_registered=user.created_at,
+            username=user.first_name + " " + user.last_name,
+        )
 
     async def change_password(
         self, user_id: str, old_password: str, new_password: str
